Create an install target that sets up USMLConfig.cmake

+++ netcdf/test/read_bathy_test.cc
@@ -335,7 +335,7 @@ BOOST_AUTO_TEST_CASE( nonglobal_database ) {
     // compare latitude axis to values read using ncdump
 
     const seq_vector& latitude = *(bathy.axis(0)) ;
-    int num_lat = latitude.size() ;
+    size_t num_lat = latitude.size() ;
     double lat1 = to_latitude( latitude(0) ) ;
     double lat2 = to_latitude( latitude(num_lat-1) ) ;
     double inc_lat = to_degrees( latitude.increment(0) ) ;
@@ -348,7 +348,7 @@ BOOST_AUTO_TEST_CASE( nonglobal_database ) {
     // compare longitude axis to values read using ncdump
 
     const seq_vector& longitude = *(bathy.axis(1)) ;
-    int num_lng = longitude.size() ;
+    size_t num_lng = longitude.size();
     double lng1 = to_degrees( longitude(0) ) ;
     double lng2 = to_degrees( longitude(num_lng-1) ) ;
     double inc_lng = to_degrees( longitude.increment(0) ) ;
